backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
APScheduler==3.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
async-timeout
asyncmy==0.2.10
attrs==25.3.0
Authlib==1.6.0
azure-cognitiveservices-search-websearch==2.0.1
azure-common==1.1.28
azure-core==1.34.0
azure-mgmt-core==1.5.0
bcrypt==3.2.0
beautifulsoup4==4.13.4
boto3==1.38.45
botocore==1.38.46
Brotli==1.1.0
bs4==0.0.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
DateTime==5.5
dill==0.4.0
distro==1.9.0
dnspython==2.7.0
doclayout_yolo==0.0.2b1
duckduckgo_search==8.0.4
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
ezdxf==1.4.2
fast-langdetect==0.2.5
fastapi
fastapi-cli
fasttext-predict==0.9.2.4
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
ftfy==6.3.1
greenlet==3.2.2
grpcio==1.67.1
h11==0.16.0
hf-xet==1.1.5
hiredis==3.2.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.1
humanfriendly==10.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain
langchain-core
langchain-community
langchain-openai
langchain-ollama
langdetect
langsmith
ldap3==2.9.1
loguru==0.7.3
lxml==6.0.0
mineru
Markdown==3.8.2
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
milvus-lite==2.5.1
minio==7.2.15
modelscope==1.27.1
more-itertools==10.7.0
mpmath==1.3.0
msrest==0.7.1
multidict==6.6.2
mysql-connector-python==9.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.6
oauthlib==3.3.1
ollama
omegaconf==2.3.0
onnxruntime==1.22.0
openai==1.93.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orjson==3.10.18
packaging==24.2
pandas==2.3.0
passlib==1.7.4
pdfminer.six==20250506
pillow==11.2.1
primp==0.15.0
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.23.0
pydantic
pydantic_validation_decorator==0.1.4
Pygments==2.19.2
PyJWT==2.10.1
pymilvus==2.5.11
pymongo==4.13.2
PyMuPDF==1.24.14
PyMySQL==1.1.0
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
docx2txt==0.8
openpyxl==3.1.2
xlrd==2.0.1
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.9
python-ulid==1.1.0
pytz==2025.2
PyYAML==6.0.2
rapid-table==1.0.5
redis==5.2.1
redis-om==0.3.5
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.6
robust-downloader==0.0.2
rsa==4.9.1
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
shapely==2.1.1
shellingham==1.5.4
simsimd==6.4.9
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sqlglot==26.6.0
sqlglotrs==0.3.14
starlette==0.36.3
stringzilla==3.12.5
sympy==1.14.0
tenacity==9.1.2
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.0
typer==0.15.3
types-cffi==1.17.0.20250523
types-pyOpenSSL==24.1.0.20240722
types-redis==4.6.0.20241004
types-setuptools==80.9.0.20250529
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
ujson==5.10.0
ultralytics==8.3.161
ultralytics-thop==2.0.14
urllib3==2.5.0
user-agents==2.2.0
uvicorn==0.29.0
uvloop==0.21.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.1
zope.interface==7.2
zstandard==0.23.0
sentence-transformers==3.0.1
tiktoken
alembic
